requirements.txt
tflite-runtime>=2.14.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
